feat:add plots for raw and processed data
